# cmd/projectmanager/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/account": {
            "post": {
                "description": "add new account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "add new account",
                "parameters": [
                    {
                        "description": "새로운 Account 등록",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gin.AddAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.AddAccountResponse"
                        }
                    }
                }
            }
        },
        "/api/account/delete/{accountId}": {
            "delete": {
                "description": "delete account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "delete account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Account 삭제",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gin.DeleteAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.DeleteAccountResponse"
                        }
                    }
                }
            }
        },
        "/api/account/detail/{accountId}": {
            "get": {
                "description": "get account info by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get account info by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.GetAccountResponse"
                        }
                    }
                }
            }
        },
...
        "/api/account/nickname/{accountId}": {
            "put": {
                "description": "update nickname of existing account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "update nickname of existing account",
                "parameters": [
                    {
                        "description": "기존 Account Nickname 수정",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gin.UpdateAccountNicknameRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.UpdateAccountNicknameResponse"
                        }
                    }
                }
            }
        },
        "/api/account/password/{accountId}": {
            "put": {
                "description": "update password of existing account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "update password of existing account",
                "parameters": [
                    {
                        "description": "기존 Account Password 수정",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gin.UpdateAccountPasswordRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.UpdateAccountPasswordResponse"
                        }
                    }
                }
            }
        },
...
        "/api/repository/image/{repositoryId}": {
            "put": {
                "description": "Create or Update Image",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create or Update Image",
                "parameters": [
                    {
                        "description": "Image 정보",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gin.CreateOrUpdateImageRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repositoryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.CreateOrUpdateImageResponse"
                        }
                    }
                }
            }
        },
        "/api/repository/images/{repositoryId}": {
            "get": {
                "description": "List Image By Repository",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List Image By Repository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repositoryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.ListImageByRepository"
                        }
                    }
                }
            }
        },
        "/api/repository/latest/{repositoryId}": {
            "get": {
                "description": "Get Latest Image by Repository",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get Latest Image by Repository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repositoryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.GetLatestImageByRepository"
                        }
                    }
                }
            }
        },
        "/internal/account/{accountId}/private": {
            "get": {
                "description": "get account info by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get account info by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.GetAccountPrivateResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gin.Account": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "registry_type": {
                    "type": "string"
                },
                "registry_url": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "gin.AccountPrivate": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "gin.AddAccountRequest": {
            "type": "object",
            "required": [
                "account_password",
                "account_username",
                "registry_type"
            ],
            "properties": {
                "account_nickname": {
                    "type": "string"
                },
                "account_password": {
                    "type": "string"
                },
                "account_username": {
                    "type": "string"
                },
                "registry_type": {
                    "type": "string"
                },
                "registry_url": {
                    "type": "string"
                }
            }
        },
        "gin.AddAccountResponse": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                }
            }
        },
...
        "gin.CreateOrUpdateImageRequest": {
            "type": "object",
            "required": [
                "tag"
            ],
            "properties": {
                "tag": {
                    "type": "string"
                }
            }
        },
        "gin.CreateOrUpdateImageResponse": {
            "type": "object",
            "properties": {
                "image_id": {
                    "type": "string"
                }
            }
        },
        "gin.DeleteAccountRequest": {
            "type": "object"
        },
        "gin.DeleteAccountResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "gin.GetAccountPrivateResponse": {
            "type": "object",
            "properties": {
                "account_private": {
                    "$ref": "#/definitions/gin.AccountPrivate"
                }
            }
        },
        "gin.GetAccountResponse": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/gin.Account"
                }
            }
        },
        "gin.GetLatestImageByRepository": {
            "type": "object",
            "properties": {
                "image": {
                    "$ref": "#/definitions/gin.Image"
                }
            }
        },
...
        "gin.Image": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "repository_id": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "gin.ListImageByRepository": {
            "type": "object",
            "properties": {
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gin.Image"
                    }
                }
            }
        },
...
        "gin.UpdateAccountNicknameRequest": {
            "type": "object",
            "required": [
                "nickname"
            ],
            "properties": {
                "nickname": {
                    "type": "string"
                }
            }
        },
        "gin.UpdateAccountNicknameResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "gin.UpdateAccountPasswordRequest": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
        "gin.UpdateAccountPasswordResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}